refactor: serve modular admin assets
