+++ admin.py
@@ -13,6 +13,7 @@ import urllib.request
 import hashlib
 import random
 from http.server import BaseHTTPRequestHandler, ThreadingHTTPServer
+from pathlib import Path
 
 import events as EV
 import places as PL
@@ -61,652 +62,8 @@ Seaking Staryu Starmie MrMime Scyther Jynx Electabuzz Magmar Pinsir Tauros Magik
 Lapras Ditto Eevee Vaporeon Jolteon Flareon Porygon Omanyte Omastar Kabuto Kabutops Aerodactyl
 Snorlax Articuno Zapdos Moltres Dratini Dragonair Dragonite Mewtwo Mew""".split()
 
-PAGE = r"""<!doctype html><html><head><meta charset="utf-8">
-<meta name="viewport" content="width=device-width,initial-scale=1">
-<title>ChucnyServer World Manager</title>
-<link rel="stylesheet" href="https://unpkg.com/leaflet@1.9.4/dist/leaflet.css"/>
-<link href="https://fonts.googleapis.com/css2?family=Plus+Jakarta+Sans:wght@400;500;600;700&display=swap" rel="stylesheet">
-<style>
- :root {
-   --bg: #f4f6f9;
-   --card: #ffffff;
-   --text: #1e293b;
-   --text-muted: #64748b;
-   --border: #e2e8f0;
-   --primary: #2563eb;
-   --primary-hover: #1d4ed8;
-   --danger: #ef4444;
-   --danger-hover: #dc2626;
-   --success: #10b981;
-   --warning: #f59e0b;
-   --radius: 12px;
-   --shadow: 0 4px 6px -1px rgba(0, 0, 0, 0.05), 0 2px 4px -2px rgba(0, 0, 0, 0.05);
- }
+ASSET_DIR = Path(__file__).with_name("admin")
 
- * { box-sizing: border-box; }
- body {
-   margin: 0;
-   background: var(--bg);
-   color: var(--text);
-   font-family: 'Plus Jakarta Sans', -apple-system, BlinkMacSystemFont, sans-serif;
-   line-height: 1.5;
-   padding-bottom: 40px;
- }
-
- header {
-   background: var(--card);
-   border-bottom: 1px solid var(--border);
-   padding: 20px 24px;
-   display: flex;
-   justify-content: space-between;
-   align-items: center;
-   box-shadow: var(--shadow);
- }
-
- .brand-title {
-   margin: 0;
-   font-size: 24px;
-   font-weight: 700;
-   background: linear-gradient(135deg, #2563eb, #3b82f6);
-   -webkit-background-clip: text;
-   -webkit-text-fill-color: transparent;
- }
-
- .status-badge {
-   display: inline-flex;
-   align-items: center;
-   gap: 6px;
-   background: #ecfdf5;
-   color: #047857;
-   font-weight: 600;
-   border-radius: 20px;
-   padding: 4px 12px;
-   font-size: 12px;
-   border: 1px solid #a7f3d0;
- }
-
- .status-dot {
-   width: 8px;
-   height: 8px;
-   background: var(--success);
-   border-radius: 50%;
-   display: inline-block;
- }
-
- .container {
-   max-width: 1200px;
-   margin: 24px auto;
-   padding: 0 16px;
-   display: grid;
-   grid-template-columns: 1fr;
-   gap: 20px;
- }
-
- .card {
-   background: var(--card);
-   border: 1px solid var(--border);
-   border-radius: var(--radius);
-   padding: 20px;
-   box-shadow: var(--shadow);
- }
-
- .card-title {
-   font-size: 18px;
-   font-weight: 600;
-   margin: 0 0 16px 0;
-   color: var(--text);
-   display: flex;
-   align-items: center;
-   justify-content: space-between;
- }
-
- .grid-2 {
-   display: grid;
-   grid-template-columns: repeat(auto-fit, minmax(350px, 1fr));
-   gap: 20px;
- }
-
- .controls-row {
-   display: flex;
-   gap: 10px;
-   flex-wrap: wrap;
-   align-items: center;
- }
-
- button, select, input {
-   background: var(--card);
-   border: 1px solid var(--border);
-   color: var(--text);
-   border-radius: 8px;
-   padding: 8px 14px;
-   font: inherit;
-   font-size: 14px;
-   outline: none;
-   transition: all 0.15s ease-in-out;
- }
-
- button {
-   font-weight: 500;
-   cursor: pointer;
- }
-
- button:hover {
-   border-color: #cbd5e1;
-   background: #f8fafc;
- }
-
- button.btn-primary {
-   background: var(--primary);
-   color: white;
-   border-color: var(--primary);
- }
-
- button.btn-primary:hover {
-   background: var(--primary-hover);
-   border-color: var(--primary-hover);
- }
-
- button.btn-danger {
-   background: #fef2f2;
-   color: var(--danger);
-   border-color: #fecaca;
- }
-
- button.btn-danger:hover {
-   background: var(--danger);
-   color: white;
- }
-
- button.on {
-   background: #eff6ff;
-   color: var(--primary);
-   border-color: #bfdbfe;
-   font-weight: 600;
- }
-
- input[type="text"], input[type="number"], select {
-   background: #f8fafc;
- }
-
- input[type="text"]:focus, input[type="number"]:focus, select:focus {
-   border-color: var(--primary);
-   background: #fff;
- }
-
- #map {
-   height: 480px;
-   width: 100%;
-   border-radius: var(--radius);
-   border: 1px solid var(--border);
-   box-shadow: var(--shadow);
- }
-
- .hint {
-   font-size: 13px;
-   color: var(--text-muted);
-   margin-top: 8px;
- }
-
- .list {
-   max-height: 250px;
-   overflow-y: auto;
-   display: flex;
-   flex-direction: column;
-   gap: 8px;
- }
-
- .row {
-   display: flex;
-   align-items: center;
-   justify-content: space-between;
-   background: #f8fafc;
-   border: 1px solid var(--border);
-   border-radius: 8px;
-   padding: 10px 14px;
-   font-size: 13px;
- }
-
- .row .t { color: var(--text); font-weight: 500; }
- .row .x { color: var(--danger); cursor: pointer; font-weight: 600; padding: 2px 6px; }
- .row .x:hover { text-decoration: underline; }
-
- .tag {
-   border-radius: 6px;
-   padding: 3px 8px;
-   font-size: 11px;
-   font-weight: 600;
-   text-transform: uppercase;
-   letter-spacing: 0.05em;
- }
- .stop { background: #e0f2fe; color: #0369a1; }
- .gym { background: #ffe4e6; color: #be123c; }
- .mon { background: #dcfce7; color: #15803d; }
-
- .empty {
-   color: var(--text-muted);
-   text-align: center;
-   padding: 20px;
-   font-size: 14px;
- }
-
- .alert-warning {
-   background: #fffbeb;
-   border: 1px solid #fde68a;
-   color: #b45309;
-   border-radius: var(--radius);
-   padding: 14px 18px;
-   font-size: 14px;
-   margin-bottom: 20px;
- }
-
- .stats-bar {
-   display: flex;
-   gap: 24px;
-   font-size: 14px;
-   color: var(--text-muted);
- }
-
- .stats-bar strong {
-   color: var(--text);
- }
-</style></head><body>
-
-<header>
-  <div>
-    <h1 class="brand-title">ChucnyServer</h1>
-    <div class="hint" style="margin:2px 0 0">World &amp; Event Management Portal</div>
-  </div>
-  <div style="text-align: right;">
-    <div class="status-badge"><span class="status-dot"></span> <span id="status">RUNNING</span></div>
-  </div>
-</header>
-
-<div class="container">
-
-  <div id="warn" class="alert-warning" style="display:none;">
-    <strong>No Gyms in your world.</strong> Random stops/gyms are OFF. Select <strong>Gym</strong> below and click on the map to add one.
-  </div>
-
-  <div class="card">
-    <div class="card-title">
-      <span>Interactive World Map</span>
-      <div class="stats-bar">
-        <span>Placed Objects: <strong id="cnt">0</strong></span>
-        <span>Details: <strong id="counts">0 stops</strong></span>
-      </div>
-    </div>
-
-    <div class="controls-row" style="margin-bottom: 14px;">
-      <button id="b-stop" class="on" onclick="setMode('stop')">PokeStop</button>
-      <button id="b-gym" onclick="setMode('gym')">Gym</button>
-      <button id="b-mon" onclick="setMode('mon')">Pokemon</button>
-      <button id="b-teleport" onclick="setMode('teleport')">Teleport Trainer</button>
-      <select id="species"></select>
-      <input id="pname" placeholder="Name (optional)" size="14">
-      <input id="pimg" placeholder="Photo path/URL" size="18">
-      <button onclick="togProc('forts')" id="b-pf">Random Stops: OFF</button>
-      <button onclick="togProc('spawns')" id="b-ps">Random Pokemon: ON</button>
-      <button class="btn-danger" onclick="clearAll()">Clear All</button>
-    </div>
-
-    <div id="map"></div>
-    <div class="hint">Click the map to place the selected element. In Teleport Trainer mode, select a trainer below first.</div>
-  </div>
-
-  <div class="grid-2">
-
-    <div class="card">
-      <div class="card-title">Quick Build Ring</div>
-      <div class="controls-row">
-        <input id="ring-n" type="number" value="8" min="1" max="24" style="width: 70px" title="Stops count">
-        <input id="ring-r" type="number" value="60" min="10" max="500" style="width: 80px" title="Radius meters">
-        <button class="btn-primary" onclick="ring()">Build Ring Around Trainer</button>
-      </div>
-      <div class="hint">Creates a neat circle of stops (and a gym) at your trainer's position.</div>
-    </div>
-
-    <div class="card">
-      <div class="card-title">Import High-Density POIs</div>
-      <div class="controls-row">
-        <input id="poi-lat" type="number" step="any" placeholder="Lat (e.g. 44.556)" style="width: 100px">
-        <input id="poi-lng" type="number" step="any" placeholder="Lng (e.g. -49.007)" style="width: 100px">
-        <input id="poi-r" type="number" value="3" min="0.1" max="50" step="0.1" style="width: 60px" title="Radius km">
-        <span style="font-size:13px; color:var(--text-muted)">km</span>
-        <input id="poi-limit" type="number" value="5000" min="100" max="10000" style="width: 75px" title="Max POIs Limit">
-        <span style="font-size:13px; color:var(--text-muted)">Limit</span>
-        <input id="poi-gym-chance" type="number" value="15" min="0" max="100" style="width: 60px" title="Gym Chance (%)">
-        <span style="font-size:13px; color:var(--text-muted)">% Gyms</span>
-        <button class="btn-primary" onclick="importPois()">Import</button>
-      </div>
-      <div class="hint" id="poihint">Imports nodes, ways, and relations from OSM including parks, art, and transit stops.</div>
-    </div>
-
-  </div>
-
-  <div class="grid-2">
-
-    <div class="card">
-      <div class="card-title">Raid Boss Management</div>
-      <div class="controls-row" style="margin-bottom: 8px;">
-        <button onclick="raidToggle()" id="b-raid">Raid: OFF</button>
-        <select id="raid-mon"></select>
-        <input id="raid-cp" type="number" value="3000" min="10" max="9999" style="width: 80px" placeholder="CP">
-        <input id="raid-name" value="raid" size="8" title="Trainer name">
-        <button class="btn-primary" onclick="raidSave()">Apply</button>
-      </div>
-      <div class="hint" id="raidhint">Spawns a raid boss defending every gym across the server.</div>
-    </div>
-
-    <div class="card">
-      <div class="card-title">Shop &amp; Storage Upgrades</div>
-      <div class="controls-row" style="margin-bottom: 12px;">
-        <span>PokeCoins: <strong id="coins" style="color:var(--warning)">0</strong></span>
-        <button onclick="buy('pokemon')" id="b-buypk">Pokemon Storage</button>
-        <button onclick="buy('items')" id="b-buyit">Item Bag</button>
-      </div>
-      <div class="controls-row" id="shopitems" style="margin-bottom: 8px;"></div>
-      <div class="hint" id="shophint">Buy items using PokeCoins earned by defending Gyms.</div>
-      <div class="hint" id="buyhint"></div>
-    </div>
-
-  </div>
-
-  <div class="grid-2">
-
-    <div class="card">
-      <div class="card-title">Player Rewards &amp; Password Reset</div>
-      <div class="controls-row" style="margin-bottom: 8px;">
-        <input id="giveuser" list="accounts" placeholder="Trainer Username (Blank = Active)" style="flex:1;">
-        <datalist id="accounts"></datalist>
-      </div>
-      <div class="controls-row" style="margin-bottom: 8px;">
-        <select id="giveitem" style="flex:1;"></select>
-        <input id="giveqty" type="number" value="20" min="1" max="999" style="width: 70px;">
-        <button class="btn-primary" onclick="give()">Give Items</button>
-      </div>
-      <div class="controls-row" style="margin-bottom: 8px;">
-        <select id="givecandy" style="flex:1;"></select>
-        <input id="givecandyqty" type="number" value="25" min="1" max="999" style="width: 70px;">
-        <button class="btn-primary" onclick="giveCandy()">Give Candy</button>
-      </div>
-      <div class="controls-row" style="margin-bottom: 8px;">
-        <input id="givedust" type="number" value="1000" min="1" max="999999" style="width: 80px;" title="Stardust">
-        <button class="btn-primary" onclick="giveDust()">Give Stardust</button>
-        <input id="givecoins" type="number" value="100" min="1" max="999999" style="width: 80px;" title="PokeCoins">
-        <button class="btn-primary" onclick="giveCoins()">Give Coins</button>
-      </div>
-      <div class="controls-row" style="margin-bottom: 8px;">
-        <input id="newpw" placeholder="New Password" style="width: 130px;">
-        <button onclick="resetPw()">Reset PW</button>
-      </div>
-      <div class="hint" id="givehint"></div>
-    </div>
-
-    <div class="card">
-      <div class="card-title">Event &amp; Global Spawn Controls</div>
-      <div class="controls-row" id="presets" style="margin-bottom: 12px;"></div>
-      <div class="controls-row">
-        <input id="ev-name" placeholder="Event Name" size="12">
-        <label style="font-size:13px; color:var(--text-muted)">Density: <input id="ev-density" type="number" min="0" max="60" style="width: 60px"></label>
-        <select id="ev-mode">
-          <option value="all">All 151</option>
-          <option value="list">From List</option>
-          <option value="single">Single Species</option>
-        </select>
-        <input id="ev-list" placeholder="1,4,7,25" size="10">
-        <label style="font-size:13px; color:var(--text-muted)">CP Range: 
-          <input id="ev-min" type="number" min="10" max="5000" style="width: 70px"> - 
-          <input id="ev-max" type="number" min="10" max="5000" style="width: 70px">
-        </label>
-        <button class="btn-primary" onclick="saveEv()">Apply Event Config</button>
-      </div>
-    </div>
-
-  </div>
-
-  <div class="grid-2">
-    <div class="card">
-      <div class="card-title">Player Nominations</div>
-      <div class="list" id="noms"><div class="empty">No nominations waiting.</div></div>
-    </div>
-
-    <div class="card">
-      <div class="card-title">Placed Objects List</div>
-      <div class="list" id="list"></div>
-    </div>
-  </div>
-
-</div>
-
-<script src="https://unpkg.com/leaflet@1.9.4/dist/leaflet.js"></script>
-<script>
-let mode='stop', map, layer, playerMarker, data={forts:[],spawns:[]};
-const $=i=>document.getElementById(i);
-const DEX=__DEX__;
-
-function setMode(m){mode=m;['stop','gym','mon','teleport'].forEach(k=>$('b-'+k).className=(k===m?'on':''));}
-function icon(color,txt){return L.divIcon({className:'',html:
- `<div style="background:${color};border:2px solid #fff;border-radius:50%;width:26px;height:26px;
-   display:flex;align-items:center;justify-content:center;font:700 11px sans-serif;color:#fff;
-   box-shadow:0 2px 5px rgba(0,0,0,0.2)">${txt}</div>`, iconSize:[26,26], iconAnchor:[13,13]});}
-
-let player={lat:0,lng:0};
-async function ring(){
-  if(!player.lat){alert('No trainer position available - open the game first or click the map.');return;}
-  await post('/api/ring',{lat:player.lat,lng:player.lng,count:+$('ring-n').value,
-    radius_m:+$('ring-r').value,gym:true});
-  load();
-}
-async function importPois(){
-  const lat = parseFloat($('poi-lat').value || player.lat);
-  const lng = parseFloat($('poi-lng').value || player.lng);
-  const r_km = parseFloat($('poi-r').value || 3);
-  const limit = parseInt($('poi-limit').value || 5000);
-  const gym_chance = parseFloat($('poi-gym-chance').value || 0) / 100.0;
-  if(!lat || !lng){alert('Please enter latitude and longitude.'); return;}
-  $('poihint').textContent = 'Fetching high-density POIs from OpenStreetMap...';
-  const r = await post('/api/fetch_pois', {lat: lat, lng: lng, radius_km: r_km, limit: limit, gym_chance: gym_chance});
-  $('poihint').textContent = r.message || ('Imported ' + r.placed + ' Objects.');
-  load();
-}
-async function saveEv(){
-  await post('/api/save',{event_name:$('ev-name').value,spawn_density:+$('ev-density').value,
-    species_mode:$('ev-mode').value,
-    species_list:$('ev-list').value.split(',').map(x=>parseInt(x.trim())).filter(x=>x>=1&&x<=151),
-    single_species:+$('species').value,min_cp:+$('ev-min').value,max_cp:+$('ev-max').value});
-  load();
-}
-async function preset(n){await post('/api/preset',{name:n});load();}
-function paintShop(coins){
-  const box=$('shopitems'); box.innerHTML='';
-  (SHOP).forEach(([sku,label,iid,cnt,price])=>{
-    const b=document.createElement('button');
-    b.textContent = label + ' (' + price + 'c)';
-    b.disabled = coins < price;
-    b.style.opacity = coins < price ? 0.45 : 1;
-    b.onclick = ()=>buyItem(sku);
-    box.appendChild(b);
-  });
-}
-async function buyItem(sku){
-  const r = await post('/api/buyitem', {sku: sku, player: $('giveuser').value});
-  $('buyhint').textContent = (r.ok?'✓ ':'✗ ') + r.message;
-  $('buyhint').style.color = r.ok ? 'var(--success)' : 'var(--danger)';
-  load();
-}
-function raidPaint(r){
-  $('b-raid').textContent = 'Raid: ' + (r.on ? 'ON' : 'OFF');
-  $('b-raid').className = r.on ? 'btn-danger' : '';
-  if(r.pokemon_id) $('raid-mon').value = r.pokemon_id;
-  if(r.cp) $('raid-cp').value = r.cp;
-  if(r.trainer) $('raid-name').value = r.trainer;
-}
-async function raidToggle(){
-  const r = await post('/api/raid', {on: $('b-raid').textContent.indexOf('ON') < 0,
-    pokemon_id:+$('raid-mon').value, cp:+$('raid-cp').value, trainer:$('raid-name').value});
-  raidPaint(r); $('raidhint').textContent = r.message; load();
-}
-async function raidSave(){
-  const r = await post('/api/raid', {pokemon_id:+$('raid-mon').value,
-    cp:+$('raid-cp').value, trainer:$('raid-name').value});
-  raidPaint(r); $('raidhint').textContent = r.message; load();
-}
-async function loadNoms(){
-  const r = await post('/api/noms', {});
-  const box = $('noms');
-  if (!r.rows || !r.rows.length){
-    box.innerHTML = '<div class="empty">No nominations waiting.</div>'; return;
-  }
-  box.innerHTML = '';
-  r.rows.forEach(function(n){
-    const d = document.createElement('div'); d.className='row';
-    d.innerHTML = '<div><span class="tag ' + (n.kind==='gym'?'gym':'stop') + '">'
-      + (n.kind==='gym'?'GYM':'STOP') + '</span> '
-      + '<b>' + n.name + '</b> &mdash; <small style="color:var(--text-muted)">by ' + n.player
-      + (n.note ? ' &middot; ' + n.note : '') + '</small></div>';
-    const no = document.createElement('span');
-    no.className='x'; no.textContent='Remove';
-    no.onclick = function(){ resolveNom(n.id,'rejected'); };
-    d.appendChild(no); box.appendChild(d);
-  });
-}
-async function resolveNom(id, status){
-  await post('/api/noms/resolve', {id: id, status: status});
-  loadNoms(); load();
-}
-function giveResult(r){
-  $('givehint').textContent = (r.ok?'✓ ':'✗ ') + r.message;
-  $('givehint').style.color = r.ok ? 'var(--success)' : 'var(--danger)';
-  load();
-}
-async function give(){
-  giveResult(await post('/api/give',{player:$('giveuser').value,
-    kind:'item', item_id:+$('giveitem').value, count:+$('giveqty').value}));
-}
-async function giveCandy(){
-  giveResult(await post('/api/give',{player:$('giveuser').value,
-    kind:'candy', pokemon_id:+$('givecandy').value, count:+$('givecandyqty').value}));
-}
-async function giveDust(){
-  giveResult(await post('/api/give',{player:$('giveuser').value,
-    kind:'stardust', count:+$('givedust').value}));
-}
-async function giveCoins(){
-  giveResult(await post('/api/give',{player:$('giveuser').value,
-    kind:'coins', count:+$('givecoins').value}));
-}
-async function resetPw(){
-  const who = $('giveuser').value.trim();
-  if(!who) return giveResult({ok:false, message:'Select a trainer first'});
-  giveResult(await post('/api/setpw', {player: who, password: $('newpw').value}));
-  $('newpw').value='';
-}
-async function buy(kind){
-  const r=await post('/api/buy',{what:kind});
-  $('shophint').textContent = (r.ok?'✓ ':'✗ ') + r.message;
-  $('shophint').style.color = r.ok ? 'var(--success)' : 'var(--danger)';
-  load();
-}
-
-async function load(){
-  const j=await (await fetch('/api/world')).json();
-  data=j.places;
-  const selected=$('giveuser').value.trim();
-  player=selected && j.teleports[selected]
-    ? {lat:j.teleports[selected][0],lng:j.teleports[selected][1]} : j.player;
-  $('cnt').textContent=data.forts.length+data.spawns.length;
-  const ngym=data.forts.filter(f=>f.kind==='gym').length;
-  const nstop=data.forts.length-ngym;
-  $('warn').style.display=(ngym===0&&!data.procedural_forts)?'block':'none';
-  $('counts').textContent=nstop+' Stops / '+ngym+' Gyms / '+data.spawns.length+' Spawns';
-  const st=j.storage||{};
-  $('coins').textContent=st.coins||0;
-  paintShop(st.coins||0);
-  $('b-buypk').textContent='Pokemon storage: '+(st.pokemon_used||0)+'/'+(st.max_pokemon||0);
-  $('b-buyit').textContent='Item bag: '+(st.items_used||0)+'/'+(st.max_items||0);
-  $('b-pf').textContent='Random Stops: '+(data.procedural_forts?'ON':'OFF');
-  $('b-ps').textContent='Random Pokemon: '+(data.procedural_spawns?'ON':'OFF');
-  $('b-pf').className=data.procedural_forts?'on':'';
-  $('b-ps').className=data.procedural_spawns?'on':'';
-  const c=j.config;
-  $('ev-name').value=c.event_name; $('ev-density').value=c.spawn_density;
-  $('ev-mode').value=c.species_mode; $('ev-list').value=(c.species_list||[]).join(',');
-  $('ev-min').value=c.min_cp; $('ev-max').value=c.max_cp;
-  if(!$('presets').dataset.done){
-    (j.presets||[]).forEach(n=>{const b=document.createElement('button');
-      b.textContent=n;b.onclick=()=>preset(n);$('presets').appendChild(b);});
-    $('presets').dataset.done='1';
-  }
-  if(!map){
-    map=L.map('map').setView([player.lat||39.19,player.lng||-96.58], player.lat?18:16);
-    L.tileLayer('https://tile.openstreetmap.org/{z}/{x}/{y}.png',{maxZoom:19,
-      attribution:'&copy; OpenStreetMap'}).addTo(map);
-    layer=L.layerGroup().addTo(map);
-    map.on('click',e=>place(e.latlng.lat,e.latlng.lng));
-  }
-  if(playerMarker) playerMarker.remove();
-  if(player.lat||player.lng) playerMarker=L.circleMarker([player.lat,player.lng],{radius:7,color:'#2563eb',
-    fillColor:'#3b82f6',fillOpacity:1}).addTo(map).bindTooltip('Trainer Position');
-  draw();
-}
-function draw(){
-  layer.clearLayers();
-  data.forts.forEach(f=>L.marker([f.lat,f.lng],{icon:icon(f.kind==='gym'?'#ef4444':'#2563eb',
-    f.kind==='gym'?'G':'S')}).addTo(layer).bindTooltip(f.name).on('click',()=>del(f.id)));
-  data.spawns.forEach(s=>L.marker([s.lat,s.lng],{icon:icon(s.pokemon_id?'#10b981':'#8b5cf6',s.pokemon_id?String(s.pokemon_id):'?')})
-    .addTo(layer).bindTooltip(s.pokemon_id?(DEX[s.pokemon_id]||('#'+s.pokemon_id)):'Random').on('click',()=>del(s.id)));
-  const rows=[...data.forts.map(f=>({id:f.id,cls:f.kind==='gym'?'gym':'stop',
-      tag:f.kind==='gym'?'GYM':'STOP',
-      t:`${f.name}${f.image?' [photo]':''} — ${f.lat.toFixed(5)}, ${f.lng.toFixed(5)}`})),
-    ...data.spawns.map(s=>({id:s.id,cls:'mon',tag:s.pokemon_id?'MON':'RANDOM',
-      t:`${s.pokemon_id?(DEX[s.pokemon_id]||('#'+s.pokemon_id)):'Random Pokemon'} — ${s.lat.toFixed(5)}, ${s.lng.toFixed(5)}`}))];
-  $('list').innerHTML = rows.length ? rows.map(r=>
-    `<div class="row"><div><span class="tag ${r.cls}">${r.tag}</span> <span class="t">${r.t}</span></div>
-     <span class="x" onclick="del('${r.id}')">Remove</span></div>`).join('')
-    : '<div class="empty">Nothing placed yet — click the map above.</div>';
-}
-async function post(u,b){return (await fetch(u,{method:'POST',headers:{'Content-Type':'application/json'},
-  body:JSON.stringify(b||{})})).json();}
-async function place(lat,lng){
-  if(mode==='teleport'){
-    const player=$('giveuser').value.trim();
-    if(!player){alert('Select a trainer before teleporting.');return;}
-    const result=await post('/api/teleport',{player,lat,lng});
-    if(!result.ok){alert(result.message);return;}
-    $('givehint').textContent=`Teleported ${result.player} to ${lat.toFixed(5)}, ${lng.toFixed(5)}`;
-    $('givehint').style.color='var(--success)';
-    return load();
-  }
-  const name=$('pname').value;
-  if(mode==='mon') await post('/api/spawn',{lat,lng,pokemon_id:+$('species').value,name});
-  else await post('/api/fort',{lat,lng,kind:mode,name,image:$('pimg').value});
-  load();
-}
-async function del(id){await post('/api/remove',{id});load();}
-async function clearAll(){if(confirm('Remove every placed object?')){await post('/api/clear',{});load();}}
-async function togProc(what){
-  const cur = what==='forts' ? data.procedural_forts : data.procedural_spawns;
-  await post('/api/procedural',{on:!cur, what}); load();
-}
-const SHOP = __SHOP__;
-(__GIVEABLE__).forEach(([id,label])=>{const o=document.createElement('option');
-  o.value=id;o.textContent=label;$('giveitem').appendChild(o);});
-DEX.forEach((n,i)=>{if(i){const o=document.createElement('option');o.value=i;
-  o.textContent=n+' candy';$('givecandy').appendChild(o);}});
-$('givecandy').value=25;
-DEX.forEach((n,i)=>{if(i){const o=document.createElement('option');o.value=i;
-  o.textContent=i+' '+n;$('raid-mon').appendChild(o);}});
-$('raid-mon').value=150;
-post('/api/raid',{}).then(raidPaint);
-post('/api/accounts',{}).then(r=>{(r.accounts||[]).forEach(n=>{
-  const o=document.createElement('option');o.value=n;$('accounts').appendChild(o);});});
-{const o=document.createElement('option');o.value=0;
- o.textContent='Random Pokemon';$('species').appendChild(o);}
-DEX.forEach((n,i)=>{if(i){const o=document.createElement('option');o.value=i;
-  o.textContent=i+' '+n;$('species').appendChild(o);}});
-$('species').value=0;
-load(); loadNoms(); setInterval(load, 15000); setInterval(loadNoms, 20000);
-</script></body></html>"""
 
 
 def fetch_overpass_pois(lat: float, lng: float, radius_m: float = 3000.0, limit: int = 10000, gym_chance: float = 0.0) -> int:
@@ -818,11 +175,22 @@ class _Handler(BaseHTTPRequestHandler):
 
     def do_GET(self):
         p = self.path.split("?")[0]
-        if p == "/":
-            return self._send(200, "text/html; charset=utf-8",
-                              PAGE.replace("__DEX__", json.dumps(DEX))
-                                  .replace("__GIVEABLE__", json.dumps(GIVEABLE))
-                                  .replace("__SHOP__", json.dumps(SHOP)))
+        assets = {
+            "/": ("index.html", "text/html; charset=utf-8"),
+            "/admin/styles.css": ("styles.css", "text/css; charset=utf-8"),
+            "/admin/app.js": ("app.js", "application/javascript; charset=utf-8"),
+        }
+        if p in assets:
+            name, ctype = assets[p]
+            try:
+                body = (ASSET_DIR / name).read_text(encoding="utf-8")
+            except OSError:
+                return self._send(404, "text/plain; charset=utf-8", "not found")
+            if name == "app.js":
+                body = (body.replace("__DEX__", json.dumps(DEX))
+                            .replace("__GIVEABLE__", json.dumps(GIVEABLE))
+                            .replace("__SHOP__", json.dumps(SHOP)))
+            return self._send(200, ctype, body)
         if p == "/api/world":
             try:
                 import rpc
