+++ admin/styles.css
@@ -1,0 +1,242 @@
+ :root {
+   --bg: #f4f6f9;
+   --card: #ffffff;
+   --text: #1e293b;
+   --text-muted: #64748b;
+   --border: #e2e8f0;
+   --primary: #2563eb;
+   --primary-hover: #1d4ed8;
+   --danger: #ef4444;
+   --danger-hover: #dc2626;
+   --success: #10b981;
+   --warning: #f59e0b;
+   --radius: 12px;
+   --shadow: 0 4px 6px -1px rgba(0, 0, 0, 0.05), 0 2px 4px -2px rgba(0, 0, 0, 0.05);
+ }
+
+ * { box-sizing: border-box; }
+ body {
+   margin: 0;
+   background: var(--bg);
+   color: var(--text);
+   font-family: 'Plus Jakarta Sans', -apple-system, BlinkMacSystemFont, sans-serif;
+   line-height: 1.5;
+   padding-bottom: 40px;
+ }
+
+ header {
+   background: var(--card);
+   border-bottom: 1px solid var(--border);
+   padding: 20px 24px;
+   display: flex;
+   justify-content: space-between;
+   align-items: center;
+   box-shadow: var(--shadow);
+ }
+
+ .brand-title {
+   margin: 0;
+   font-size: 24px;
+   font-weight: 700;
+   background: linear-gradient(135deg, #2563eb, #3b82f6);
+   -webkit-background-clip: text;
+   -webkit-text-fill-color: transparent;
+ }
+
+ .status-badge {
+   display: inline-flex;
+   align-items: center;
+   gap: 6px;
+   background: #ecfdf5;
+   color: #047857;
+   font-weight: 600;
+   border-radius: 20px;
+   padding: 4px 12px;
+   font-size: 12px;
+   border: 1px solid #a7f3d0;
+ }
+
+ .status-dot {
+   width: 8px;
+   height: 8px;
+   background: var(--success);
+   border-radius: 50%;
+   display: inline-block;
+ }
+
+ .container {
+   max-width: 1200px;
+   margin: 24px auto;
+   padding: 0 16px;
+   display: grid;
+   grid-template-columns: 1fr;
+   gap: 20px;
+ }
+
+ .card {
+   background: var(--card);
+   border: 1px solid var(--border);
+   border-radius: var(--radius);
+   padding: 20px;
+   box-shadow: var(--shadow);
+ }
+
+ .card-title {
+   font-size: 18px;
+   font-weight: 600;
+   margin: 0 0 16px 0;
+   color: var(--text);
+   display: flex;
+   align-items: center;
+   justify-content: space-between;
+ }
+
+ .grid-2 {
+   display: grid;
+   grid-template-columns: repeat(auto-fit, minmax(350px, 1fr));
+   gap: 20px;
+ }
+
+ .controls-row {
+   display: flex;
+   gap: 10px;
+   flex-wrap: wrap;
+   align-items: center;
+ }
+
+ button, select, input {
+   background: var(--card);
+   border: 1px solid var(--border);
+   color: var(--text);
+   border-radius: 8px;
+   padding: 8px 14px;
+   font: inherit;
+   font-size: 14px;
+   outline: none;
+   transition: all 0.15s ease-in-out;
+ }
+
+ button {
+   font-weight: 500;
+   cursor: pointer;
+ }
+
+ button:hover {
+   border-color: #cbd5e1;
+   background: #f8fafc;
+ }
+
+ button.btn-primary {
+   background: var(--primary);
+   color: white;
+   border-color: var(--primary);
+ }
+
+ button.btn-primary:hover {
+   background: var(--primary-hover);
+   border-color: var(--primary-hover);
+ }
+
+ button.btn-danger {
+   background: #fef2f2;
+   color: var(--danger);
+   border-color: #fecaca;
+ }
+
+ button.btn-danger:hover {
+   background: var(--danger);
+   color: white;
+ }
+
+ button.on {
+   background: #eff6ff;
+   color: var(--primary);
+   border-color: #bfdbfe;
+   font-weight: 600;
+ }
+
+ input[type="text"], input[type="number"], select {
+   background: #f8fafc;
+ }
+
+ input[type="text"]:focus, input[type="number"]:focus, select:focus {
+   border-color: var(--primary);
+   background: #fff;
+ }
+
+ #map {
+   height: 480px;
+   width: 100%;
+   border-radius: var(--radius);
+   border: 1px solid var(--border);
+   box-shadow: var(--shadow);
+ }
+
+ .hint {
+   font-size: 13px;
+   color: var(--text-muted);
+   margin-top: 8px;
+ }
+
+ .list {
+   max-height: 250px;
+   overflow-y: auto;
+   display: flex;
+   flex-direction: column;
+   gap: 8px;
+ }
+
+ .row {
+   display: flex;
+   align-items: center;
+   justify-content: space-between;
+   background: #f8fafc;
+   border: 1px solid var(--border);
+   border-radius: 8px;
+   padding: 10px 14px;
+   font-size: 13px;
+ }
+
+ .row .t { color: var(--text); font-weight: 500; }
+ .row .x { color: var(--danger); cursor: pointer; font-weight: 600; padding: 2px 6px; }
+ .row .x:hover { text-decoration: underline; }
+
+ .tag {
+   border-radius: 6px;
+   padding: 3px 8px;
+   font-size: 11px;
+   font-weight: 600;
+   text-transform: uppercase;
+   letter-spacing: 0.05em;
+ }
+ .stop { background: #e0f2fe; color: #0369a1; }
+ .gym { background: #ffe4e6; color: #be123c; }
+ .mon { background: #dcfce7; color: #15803d; }
+
+ .empty {
+   color: var(--text-muted);
+   text-align: center;
+   padding: 20px;
+   font-size: 14px;
+ }
+
+ .alert-warning {
+   background: #fffbeb;
+   border: 1px solid #fde68a;
+   color: #b45309;
+   border-radius: var(--radius);
+   padding: 14px 18px;
+   font-size: 14px;
+   margin-bottom: 20px;
+ }
+
+ .stats-bar {
+   display: flex;
+   gap: 24px;
+   font-size: 14px;
+   color: var(--text-muted);
+ }
+
+ .stats-bar strong {
+   color: var(--text);
+ }
